Move executor to utils
